requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
Django==1.11.7
django-flat-responsive==2.0
django-telegrambot==1.0.1
djangorestframework==3.3.2
future==0.16.0
kombu==4.1.0
olefile==0.44
Pillow==4.3.0
python-telegram-bot==8.1.1
pytz==2017.3
vine==1.1.4
